+++ src/main/java/insane96mcp/enhancedai/modules/spider/webber/TemporaryCobwebTask.java
@@ -3,20 +3,22 @@ package insane96mcp.enhancedai.modules.spider.webber;
 import insane96mcp.insanelib.world.scheduled.ScheduledTickTask;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
 
 public class TemporaryCobwebTask extends ScheduledTickTask {
 
-	Level world;
+	Level level;
 	BlockPos pos;
 
-	public TemporaryCobwebTask(int tickDelay, Level world, BlockPos pos) {
+	public TemporaryCobwebTask(int tickDelay, Level level, BlockPos pos) {
 		super(tickDelay);
-		this.world = world;
+		this.level = level;
 		this.pos = pos;
 	}
 
 	@Override
 	public void run() {
-		this.world.destroyBlock(pos, false);
+        if (this.level.getBlockState(pos).is(Blocks.COBWEB))
+		    this.level.destroyBlock(pos, false);
 	}
 }